==> 2.1.all.peaks.pie.R <==
#2.1 all peaks pie plot
library(org.Ss.eg.db)
library(ChIPseeker) #ChIPseeker v1.26.2 
library(GenomicRanges)
library(org.Sscrofa.eg.db) # created by  Penny using custome script  0.generate.OrgDb.geneannotation.R

common.merged.peaks.6798.6800 <- readRDS("path.../common.merged.peaks.6798.6800.filtered.rds") #110444

library(biomaRt)
library(GenomicFeatures)
sscrofaEnsembl <- makeTxDbFromBiomart(biomart="ENSEMBL_MART_ENSEMBL",dataset="sscrofa_gene_ensembl")#this script only works in R4.1.1, does not work in R 4.0.3

peakAnno <- annotatePeak(common.merged.peaks.6798.6800, tssRegion=c(-3000, 3000),
                         TxDb=sscrofaEnsembl, annoDb="org.Sscrofa.eg.db")
# Annotated peaks generated by ChIPseeker
# 110444/110444  peaks were annotated
# Genomic Annotation Summary:
#   Feature   Frequency
# 9    Promoter (<=1kb) 15.34261707
# 10   Promoter (1-2kb)  3.47687516
# 11   Promoter (2-3kb)  3.92506610
# 4              5' UTR  0.96066785
# 3              3' UTR  3.07667234
# 1            1st Exon  0.03169027
# 7          Other Exon  5.23885408
# 2          1st Intron 11.02368621
# 8        Other Intron 28.96309442
# 6  Downstream (<=300)  0.16116765
# 5   Distal Intergenic 27.79960885
# 
# 15.34261707+3.47687516+3.92506610=22.74456

plotAnnoPie(peakAnno)
